set datatypes defined with long back to int because long=8 bits on 64 bits archs and using long is wrong for bmp header sizes on 64 bits archs

+++ src/gfx/aux_texture.cpp
@@ -36,8 +36,8 @@ using std::string;
 
 
 #ifndef WIN32
-typedef unsigned long DWORD;
-typedef long  LONG;
+typedef unsigned int DWORD;
+typedef int  LONG;
 typedef unsigned short WORD;
 typedef unsigned char BYTE;
 /**
